+++ apps/ui-dev/CHANGELOG.md
@@ -1,5 +1,12 @@
 # ui-dev
 
+## 1.0.29
+
+### Patch Changes
+
+- update ghost, blank documentation
+  add selected status for card component
+
 ## 1.0.28
 
 ### Patch Changes